task: initial draft of curated experiments
I realized how much of a disaster my Github repos have become recently.
Not representing the code I write for work. I need to rectify that, and
as a step forward into figuring out how and what I want to show, I
decided to start with a page here that I can update and add new projects
as I go.